Splitting the \Core\Error controller in a real App Controller and an Error Messages generator; also several another fixes.
