+++ system/Core/Error.php
@@ -10,47 +10,11 @@
 
 namespace Smvc\Core;
 
-use Smvc\Core\Controller;
-use Smvc\Core\View;
-
 /**
  * Error class to generate 404 pages.
  */
-class Error extends Controller
+class Error
 {
-    /**
-     * $error holder.
-     *
-     * @var string
-     */
-    private $error = null;
-
-    /**
-     * Save error to $this->error.
-     *
-     * @param string $error
-     */
-    public function __construct($error = null)
-    {
-        parent::__construct();
-        $this->error = $error;
-    }
-
-    /**
-     * Load a 404 page with the error message.
-     */
-    public function index($error = null)
-    {
-        header("HTTP/1.0 404 Not Found");
-
-        $data['title'] = '404';
-        $data['error'] = $error ? $error : $this->error;
-
-        View::renderTemplate('header', $data);
-        View::render('error/404', $data);
-        View::renderTemplate('footer', $data);
-    }
-
     /**
      * Display errors.
      *
