+++ system/Core/Logger.php
@@ -10,7 +10,7 @@
 
 namespace Smvc\Core;
 
-use Smvc\Helpers\PhpMailer\Mail;
+use Smvc\Helpers\PhpMailer as Mailer;
 
 /**
  * Record and email/display errors or a custom error message.
@@ -190,11 +190,13 @@ class Logger
     public static function sendEmail($message)
     {
         if (self::$emailError == true) {
-            $mail = new Mail();
+            $mail = new Mailer();
+            
             $mail->setFrom(SITEEMAIL);
             $mail->addAddress(SITEEMAIL);
             $mail->subject('New error on '.SITETITLE);
             $mail->body($message);
+            
             $mail->send();
         }
     }
